+++ CHANGELOG.md
@@ -23,6 +23,9 @@ extension methods to `AssetManager` to provide an alternative to equivalent util
   - **`GdxIntArray`**: `com.badlogic.gdx.utils.IntArray`
   - **`GdxFloatArray`**: `com.badlogic.gdx.utils.FloatArray`
   - **`GdxBooleanArray`**: `com.badlogic.gdx.utils.BooleanArray`
+  - **`GdxCharArray`**: `com.badlogic.gdx.utils.CharArray`
+  - **`GdxLongArray`**: `com.badlogic.gdx.utils.LongArray`
+  - **`GdxShortArray`**: `com.badlogic.gdx.utils.ShortArray`
   - **`GdxSet`**: `com.badlogic.gdx.utils.ObjectSet`
   - **`GdxMap`**: `com.badlogic.gdx.utils.ObjectMap`
   - **`GdxList`**: `ktx.collections.PooledList`